draft:  implement an option --search for online search and drop searching on the internet as a default behavior.
