+++ src/Microsoft.TemplateEngine.Cli/CommandParsing/CommandParserSupport.cs
@@ -143,7 +143,8 @@ namespace Microsoft.TemplateEngine.Cli.CommandParsing
                     Create.Option("--update-apply", LocalizableStrings.UpdateApplyCommandHelp, Accept.NoArguments()),
                     Create.Option("--columns", LocalizableStrings.OptionDescriptionColumns, Accept.ExactlyOneArgument().With(LocalizableStrings.OptionDescriptionColumns, "COLUMNS_LIST")),
                     Create.Option("--columns-all", LocalizableStrings.OptionDescriptionColumnsAll, Accept.NoArguments()),
-                    Create.Option("--author", LocalizableStrings.OptionDescriptionAuthorFilter, Accept.ExactlyOneArgument().With(LocalizableStrings.OptionDescriptionColumns, "AUTHOR"))
+                    Create.Option("--author", LocalizableStrings.OptionDescriptionAuthorFilter, Accept.ExactlyOneArgument().With(LocalizableStrings.OptionDescriptionColumns, "AUTHOR")),
+                    Create.Option("--search", LocalizableStrings.SearchOnlineCommandHelp, Accept.NoArguments())
                 };
             }
         }
